Add ALPN support
ALPN is basically a replacement for NPN that is going to be used by HTTP/2.0
and is likely to get IETF blessing. It has been added to OpenSSL branch
targetting 1.0.2 release and NPN is planned to be removed some time after that.

This change adds support for ALPN, largely based on existing NPN support,
along with a test. It works with snapshots of OpenSSL 1.0.2 and patched
Net::SSLeay with the following patch:
http://alioth.debian.org/tracker/index.php?func=detail&aid=314500&group_id=30843&atid=412216